requirements.txt
fastapi
uvicorn[standard]
motor
pymongo
python-dotenv
pydantic[email]
passlib[bcrypt]
